jsonpath: refactor walkAST and reorganize implementation
This commit refactors walkAST to a Validate method within the Path
interface to implement for all path types.

Release note: None
